Update GetAndroidApiLevelResponse to use Long for apiLevel and adjust AutomatorServer implementation accordingly

+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/AutomatorServer.kt
@@ -372,6 +372,6 @@ class AutomatorServer(private val automation: Automator) : NativeAutomatorServer
     }
 
     override fun getAndroidApiLevel(): GetAndroidApiLevelResponse {
-        return GetAndroidApiLevelResponse(Build.VERSION.SDK_INT)
+        return GetAndroidApiLevelResponse(Build.VERSION.SDK_INT.toLong())
     }
 }
