+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/contracts/Contracts.kt
@@ -680,7 +680,7 @@ class Contracts {
   )
 
   data class GetAndroidApiLevelResponse (
-    val apiLevel: Int
+    val apiLevel: Long
   )
 
 }
